[tasks] Share the board status logic between Todos and Person
Extract the boardStatuses computed and getBoardStatusesByProduction,
duplicated between the two pages, into a useBoardStatuses composable
parameterized by the production list and the selected production. Each
page drops its now-unused taskStatuses and getProductionTaskStatuses
getters and its sortTaskStatuses import.
